requirements.txt
# Core
langchain==0.3.26
langchain-core==0.3.80
langchain-text-splitters==0.3.11
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-huggingface==0.3.1
# LLM / embeddings
sentence-transformers==4.1.0
pydantic>=2.0.0,<3.0.0
pypdf==5.6.1

# Web / app
flask==3.1.1
python-dotenv==1.1.0

# Local package
-e .
